requirements.txt
django
Pillow
tzlocal
python-dateutil